+++ tests/test_config.py
@@ -2,8 +2,9 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from rich_click import group, rich_config, RichHelpConfiguration, RichContext
+from rich_click import group, command, rich_config, RichHelpConfiguration, RichContext
 from rich_click._compat_click import CLICK_IS_BEFORE_VERSION_8X
+import rich_click.rich_click as rc
 
 if CLICK_IS_BEFORE_VERSION_8X:
     pytest.skip(reason="rich_config not supported for click < 8.", allow_module_level=True)
@@ -51,3 +52,58 @@ def test_basic_config_for_group() -> None:
         with subcommand3.make_context("pytest-example", [], parent=ctx) as sub_ctx:
             assert isinstance(sub_ctx, RichContext)
             assert sub_ctx.help_config.style_option == "overwriting-parent-value"
+
+
+def test_global_config_equal_config_defaults() -> None:
+    config1 = RichHelpConfiguration()
+    config2 = RichHelpConfiguration.load_from_globals(rc)
+    for k in {*config1.__dict__.keys(), *config2.__dict__.keys()}:
+        if k == "highlighter":
+            continue
+        v1 = config1.__dict__[k]
+        v2 = config2.__dict__[k]
+        assert v1 == v2, k
+
+
+def test_config_from_globals_behavior() -> None:
+    original_style_option_value = rc.STYLE_OPTION
+    rc.STYLE_OPTION = "new-value"
+
+    @command()
+    def cli1() -> None:
+        pass
+
+    with cli1.make_context("pytest-example", []) as ctx:
+        if TYPE_CHECKING:
+            assert isinstance(ctx, RichContext)
+        assert ctx.help_config.style_option == "new-value"
+
+    @command()
+    @rich_config(help_config=RichHelpConfiguration())
+    def cli2() -> None:
+        pass
+
+    with cli2.make_context("pytest-example", []) as ctx:
+        if TYPE_CHECKING:
+            assert isinstance(ctx, RichContext)
+        assert ctx.help_config.style_option == original_style_option_value
+
+    @command()
+    @rich_config(help_config=RichHelpConfiguration.load_from_globals())
+    def cli3() -> None:
+        pass
+
+    with cli3.make_context("pytest-example", []) as ctx:
+        if TYPE_CHECKING:
+            assert isinstance(ctx, RichContext)
+        assert ctx.help_config.style_option == "new-value"
+
+    @command()
+    @rich_config(help_config={})
+    def cli4() -> None:
+        pass
+
+    with cli4.make_context("pytest-example", []) as ctx:
+        if TYPE_CHECKING:
+            assert isinstance(ctx, RichContext)
+        assert ctx.help_config.style_option == "new-value"